+++ models/db/Consultation.php
@@ -379,7 +379,8 @@ class Consultation extends ActiveRecord
             IMotion::STATUS_DELETED,
             IMotion::STATUS_UNCONFIRMED,
             IMotion::STATUS_DRAFT,
-            IMotion::STATUS_COLLECTING_SUPPORTERS
+            IMotion::STATUS_COLLECTING_SUPPORTERS,
+            IMotion::STATUS_DRAFT_ADMIN,
         ];
         if (!$this->getSettings()->screeningMotionsShown) {
             $invisible[] = IMotion::STATUS_SUBMITTED_UNSCREENED;
@@ -401,6 +402,7 @@ class Consultation extends ActiveRecord
             IMotion::STATUS_DELETED,
             IMotion::STATUS_UNCONFIRMED,
             IMotion::STATUS_DRAFT,
+            IMotion::STATUS_DRAFT_ADMIN,
         ];
         return $invisible;
     }