+++ models/db/Motion.php
@@ -440,6 +440,7 @@ class Motion extends IMotion implements IRSSItem
                 static::STATUS_DRAFT,
                 static::STATUS_COLLECTING_SUPPORTERS,
                 static::STATUS_SUBMITTED_UNSCREENED,
+                static::STATUS_DRAFT_ADMIN,
             ];
             if (in_array($this->status, $notAmendableStati)) {
                 if ($throwExceptions) {